businesslogic created

+++ src/main/java/net/javaguides/springboot_transaction_demo/entity/Payment.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Table(name="payment")
+@Table(name = "payment")
 public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,5 +29,5 @@ public class Payment {
     @Column(nullable = false)
     private int cvc;
     @Column(nullable = false)
-    private int orderId;
+    private Long orderId;
 }
